requirements.txt
streamlit
torch
torchvision
ultralytics
opencv-python-headless
pillow
numpy
